make a file for ShardedDataSource, move Json/HF/Jsonl there
